% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaSensCurve.R
\name{AlphaSensAnalysis}
\alias{AlphaSensAnalysis}
\title{Sensitivity alpha-curve associated to MLP function}
\usage{
AlphaSensAnalysis(sens, tol = NULL, max_alpha = 1000)
}
\arguments{
\item{sens}{sensitivity object returned by \code{\link[NeuralSens]{SensAnalysisMLP}}}

\item{tol}{difference between M_alpha and maximum sensitivity of the sensitivity of each input variable}

\item{max_alpha}{maximum alpha value to analyze}
}
\value{
alpha-curves of the MLP function
}
\description{
Obtain sensitivity alpha-curves associated to MLP function obtained from
the sensitivities returned by \code{\link[NeuralSens]{SensAnalysisMLP}}.
}
