% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DAILY_DEMAND_TR}
\alias{DAILY_DEMAND_TR}
\title{Data frame with 4 variables}
\format{A data frame with 1980 rows and 4 variables:
\describe{
  \item{fecha}{date of the measure}
  \item{DEM}{electrical demand}
  \item{WD}{working day coefficient}
  \item{TEMP}{weather temperature}
}}
\description{
Training dataset with values of temperature and working day to predict electrical demand
}
\author{
Jose Portela Gonzalez
}
\keyword{data}
