% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DAILY_DEMAND_TV}
\alias{DAILY_DEMAND_TV}
\title{Data frame with 3 variables}
\format{A data frame with 7 rows and 3 variables:
\describe{
  \item{fecha}{date of the measure}
  \item{WD}{working day coefficient}
  \item{TEMP}{weather temperature}
}}
\description{
Validation dataset with values of temperature and working day to predict electrical demand
}
\author{
Jose Portela Gonzalez
}
\keyword{data}
