% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActFunc.R
\name{DerActFunc}
\alias{DerActFunc}
\title{Derivate activation function of neuron}
\usage{
DerActFunc(type = "sigmoid", ...)
}
\arguments{
\item{type}{\code{character} name of the activation function}

\item{...}{extra arguments needed to calculate the functions}
}
\value{
\code{numeric} output of the neuron
}
\description{
Evaluate derivate of activation function of a neuron
}
