% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActFunc.R
\name{DerLossFunc}
\alias{DerLossFunc}
\title{Derivative of loss function}
\usage{
DerLossFunc(type = "RMSE", ...)
}
\arguments{
\item{type}{\code{character} name of the loss function}

\item{...}{extra arguments needed to calculate the functions}
}
\value{
\code{numeric} output of the neuron
}
\description{
Evaluate derivative of loss function of a neuron
}
\examples{
lossFunction <- DerLossFunc("RMSE")
}
