% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HessMLP.R
\name{HessToSensMLP}
\alias{HessToSensMLP}
\title{Convert a HessMLP to a SensMLP object}
\usage{
HessToSensMLP(x)
}
\arguments{
\item{x}{\code{HessMLP} object}
}
\value{
\code{SensMLP} object
}
\description{
Auxiliary function to turn a HessMLP object to a SensMLP object in order to use the plot-related functions
associated with SensMLP
}
