% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HessMLP.R
\name{is.HessMLP}
\alias{is.HessMLP}
\title{Check if object is of class \code{HessMLP}}
\usage{
is.HessMLP(object)
}
\arguments{
\item{object}{\code{HessMLP} object}
}
\value{
\code{TRUE} if \code{object} is a \code{HessMLP} object
}
\description{
Check if object is of class \code{HessMLP}
}
