% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JerkMLP.R
\name{is.JerkMLP}
\alias{is.JerkMLP}
\title{Check if object is of class \code{JerkMLP}}
\usage{
is.JerkMLP(object)
}
\arguments{
\item{object}{\code{JerkMLP} object}
}
\value{
\code{TRUE} if \code{object} is a \code{JerkMLP} object
}
\description{
Check if object is of class \code{JerkMLP}
}
