% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensMLP.R
\name{is.SensMLP}
\alias{is.SensMLP}
\title{Check if object is of class \code{SensMLP}}
\usage{
is.SensMLP(object)
}
\arguments{
\item{object}{\code{SensMLP} object}
}
\value{
\code{TRUE} if \code{object} is a \code{SensMLP} object
}
\description{
Check if object is of class \code{SensMLP}
}
\references{
Pizarroso J, Portela J, Muñoz A (2022). NeuralSens: Sensitivity Analysis of
Neural Networks. Journal of Statistical Software, 102(7), 1-36.
}
