% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JerkMLP.R
\name{JerkMLP}
\alias{JerkMLP}
\title{Constructor of the JerkMLP Class}
\usage{
JerkMLP(
  sens = list(),
  raw_sens = list(),
  mlp_struct = numeric(),
  trData = data.frame(),
  coefnames = character(),
  output_name = character()
)
}
\arguments{
\item{sens}{\code{list} of sensitivity measures, one \code{list} per output neuron}

\item{raw_sens}{\code{list} of sensitivities, one \code{array} per output neuron}

\item{mlp_struct}{\code{numeric} vector describing the structur of the MLP model}

\item{trData}{\code{data.frame} with the data used to calculate the sensitivities}

\item{coefnames}{\code{character} vector with the name of the predictor(s)}

\item{output_name}{\code{character} vector with the name of the output(s)}
}
\value{
\code{JerkMLP} object
}
\description{
Create an object of JerkMLP class
}
