% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActFunc.R
\name{LossFunc}
\alias{LossFunc}
\title{Loss function}
\usage{
LossFunc(type = "RMSE", ...)
}
\arguments{
\item{type}{\code{character} name of the loss function}

\item{...}{extra arguments needed to calculate the functions}
}
\value{
\code{list} containing the \code{numeric} error of the neural network
and a \code{numeric} vector to codify if it is a regression loss (0), a
binary classification loss (1) or a multiclass classification loss (2)
}
\description{
Evaluate loss function of a neuron
}
\examples{
lossFunction <- LossFunc("RMSE")
}
