% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensMLP.R
\name{SensMLP}
\alias{SensMLP}
\title{Constructor of the SensMLP Class}
\usage{
SensMLP(
  sens = list(),
  raw_sens = list(),
  mlp_struct = numeric(),
  trData = data.frame(),
  coefnames = character(),
  output_name = character(),
  cv = NULL,
  boot = NULL,
  boot.alpha = NULL
)
}
\arguments{
\item{sens}{\code{list} of sensitivity measures, one \code{data.frame} per output neuron}

\item{raw_sens}{\code{list} of sensitivities, one \code{matrix} per output neuron}

\item{mlp_struct}{\code{numeric} vector describing the structur of the MLP model}

\item{trData}{\code{data.frame} with the data used to calculate the sensitivities}

\item{coefnames}{\code{character} vector with the name of the predictor(s)}

\item{output_name}{\code{character} vector with the name of the output(s)}

\item{cv}{\code{list} list with critical values of significance for std and mean square.}

\item{boot}{\code{array} bootstrapped sensitivity measures.}

\item{boot.alpha}{\code{array} significance level.
Defaults to \code{NULL}. Only available for analyzed \code{caret::train} models.}
}
\value{
\code{SensMLP} object
}
\description{
Create an object of SensMLP class
}
\references{
Pizarroso J, Portela J, Muñoz A (2022). NeuralSens: Sensitivity Analysis of
Neural Networks. Journal of Statistical Software, 102(7), 1-36.
}
