% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{syntheticdata}
\alias{syntheticdata}
\title{Data frame with 15 variables}
\format{A data frame with 1000 rows and 15 variables:
\describe{
  \item{P1x1}{Input 1 of the subset 1 (numeric)}
  \item{P1x2}{Input 2 of the subset 1 (numeric)}
  \item{P1x3}{Input 3 of the subset 1 (numeric)}
  \item{P1y}{Output of the subset 1 (numeric)}
  \item{P2x1}{Input 1 of the subset 2 (numeric)}
  \item{P2x2}{Input 2 of the subset 2 (numeric)}
  \item{P2x3}{Input 3 of the subset 2 (numeric)}
  \item{P2y}{Output of the subset 2 (character)}
  \item{P3x1}{Input 1 of the subset 3 (numeric)}
  \item{P3x2}{Input 2 of the subset 3 (numeric)}
  \item{P3x3}{Input 3 of the subset 3 (numeric)}
  \item{P3y}{Output of the subset 3 (numeric)}
  \item{P4x1}{Input 1 of the subset 3 (numeric)}
  \item{P4x2}{Input 2 of the subset 3 (character)}
  \item{P4y}{Output of the subset 3 (numeric)}
}}
\description{
Dataset composed of several inputs with its corresponding outputs to test the
functionality of the package for regression
}
\author{
Jose Portela Gonzalez
}
\keyword{data}
