% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DAILY_DEMAND_TR}
\alias{DAILY_DEMAND_TR}
\title{Data frame with 4 variables}
\format{
A data frame with 1980 rows and 4 variables:
\describe{
  \item{DATE}{date of the measure}
  \item{DEM}{electrical demand}
  \item{WD}{Working Day: index which express how much work is made that day}
  \item{TEMP}{weather temperature}
}
}
\description{
Training dataset with values of temperature and working day to predict electrical demand
}
\references{
Pizarroso J, Portela J, Muñoz A (2022). NeuralSens: Sensitivity Analysis of
Neural Networks. Journal of Statistical Software, 102(7), 1-36.
}
\author{
Jose Portela Gonzalez
}
\keyword{data}
