% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralSens-package.R
\docType{package}
\name{NeuralSens}
\alias{NeuralSens}
\alias{NeuralSens-package}
\title{NeuralSens: Sensibility analysis to NeuralNets}
\description{
Visualization and analysis tools to aid in the interpretation of
neural network models.
}
