% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralSens-package.R
\docType{package}
\name{NeuralSens}
\alias{NeuralSens}
\title{NeuralSens: Sensitivity Analysis of Neural Networks}
\description{
Visualization and analysis tools to aid in the interpretation of
neural network models.
}
