% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensMLP.R
\name{SensMLP}
\alias{SensMLP}
\title{Constructor of the SensMLP Class}
\usage{
SensMLP(
  sens = list(),
  raw_sens = list(),
  mlp_struct = numeric(),
  trData = data.frame(),
  coefnames = character(),
  output_name = character()
)
}
\arguments{
\item{sens}{\code{list} of sensitivity measures, one \code{data.frame} per output neuron}

\item{raw_sens}{\code{list} of sensitivities, one \code{matrix} per output neuron}

\item{mlp_struct}{\code{numeric} vector describing the structur of the MLP model}

\item{trData}{\code{data.frame} with the data used to calculate the sensitivities}

\item{coefnames}{\code{character} vector with the name of the predictor(s)}

\item{output_name}{\code{character} vector with the name of the output(s)}
}
\value{
\code{SensMLP} object
}
\description{
Create an object of SensMLP class
}
