% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{syntheticdata}
\alias{syntheticdata}
\title{List of 4 dataframes to test the functions with different variables types}
\format{
list of 4 data.frames with 4 columns for 3 inputs and one output:
\describe{
  \item{RegOutNumInp}{data.frame}
  \itemize{
      \item{X1} {Input 1 of the subset 1 (numeric)}
      \item{X2} {Input 2 of the subset 1 (numeric)}
      \item{X3} {Input 3 of the subset 1 (numeric)}
      \item{Y} {Output of the subset 1 (numeric)}
  }
  \item{ClsOutNumInp}{data.frame}
  \itemize{
      \item{X1} {Input 1 of the subset 2 (numeric)}
      \item{X2} {Input 2 of the subset 2 (numeric)}
      \item{X3} {Input 3 of the subset 2 (numeric)}
      \item{Y} {Output of the subset 2 (character)}
  }
  \item{ClsOutClsInp}{data.frame}
  \itemize{
      \item{X1} {Input 1 of the subset 3 (character)}
      \item{X2} {Input 2 of the subset 3 (numeric)}
      \item{X3} {Input 3 of the subset 3 (numeric)}
      \item{Y} {Output of the subset 3 (character)}
  }
  \item{ClsOutClsInp}{data.frame}
  \itemize{
      \item{X1} {Input 1 of the subset 4 (numeric)}
      \item{X2} {Input 2 of the subset 4 (character)}
      \item{X3} {Input 3 of the subset 4 (numeric)}
      \item{Y} {Output of the subset 4 (numeric)}
  }
}
}
\description{
List of 4 dataframes to test the functions with different
variables types (numeric and character output and inputs)
}
\author{
Jose Portela Gonzalez
}
\keyword{data}
